backend_all/backendAdmin/requirements.txt

Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1

mypy_extensions==1.1.0
mysql-connector-python==9.4.0

neo4j==5.28.2
numpy
opencv-python==4.8.0.76
py2neo==2021.2.4
PyJWT==2.10.1
PyMySQL==1.1.2
redis==7.0.1
# torch==2.3.1


